[#1096] Change test from circle.yml to .travis.yml
