+++ .travis.yml
@@ -1,0 +1,9 @@
+language: java
+
+sudo: required
+
+services:
+  - docker
+
+script:
+  - ./ci/build.sh
